Update DTS
Add GPIOs

Outputs are set as LEDs(!) so there is easy access by label in sysfs.
Inputs appear labelled correctly when using gpiod (and can be read
with whe gpioget command).
Alas, trying to use gpioset with the outputs doesn't seem to work
but setting using /sys/class/leds does work.
So... full control is sort of available, if inconvenient.

Actually none of this is used for the SMAP programming interface,
which just memory-maps the GPIO register space for speed.